requirements.txt
# Learning
torchmetrics
open3d
pandas
wandb
pytorch-lightning>=2.0.2
hydra-core
scipy==1.7.0
scikit-image
matplotlib
zarr
numcodecs
tqdm
dask
numba
setuptools==59.5.0
mlflow
boto3
# Manipulation
autolab-core
loguru
pybullet==3.2.5
harvesters
pymongo
py_cli_interaction
gitpython
libigl
pyrfuniverse==0.10.1
easydict
learn
rich
plyer
pyclipper
transforms3d
pyquaternion
fastapi>=0.97.0
uvicorn>=0.22.0
dearpygui
python-statemachine[diagrams]
# Management
vyper-config>=1.1.1
loguru>=0.7.0
click>=8.1.3
minio>=7.1.15
pika>=1.3.2
requests>=2.31.0
openapi-python-client>=0.14.1
websockets>=11.0.3
